refactor: rename useHotkey import and move implementation to useHot.ts
