+++ src/components/sounds/main/grid.tsx
@@ -4,7 +4,7 @@ import { SoundButton } from '@/components/sounds/button'
 import { useAudio } from '@/context/audio'
 import { useSettings } from '@/context/setting'
 import { useSounds } from '@/context/sounds'
-import { useHotkey } from '@/hooks/useHotkey'
+import { useHotkey } from '@/hooks/useHot'
 
 interface SoundGridProps {
   type: 'sound' | 'music'
